:sparkles: 接口内容更改

+++ backend/src/cognicube_backend/schemas/message.py
@@ -5,7 +5,7 @@ from typing import List
 class Message(BaseModel):
     text: str
     reply_to: int | None = None
-    timestamp: int | None = None
+    timestamp: float | None = None
     who: str
     message_id: int | None = None
